Fix Vue Proxy and raw mismatch
While the proposal/candidate proxies were marked as raw, the
proxies were still added to the arrays. Vue 3.5.14 introduced
a fix so a proxy marked as raw returns the raw object. This
happened on array access, so the "nextFocussedProposal/Candidate"
was the raw object but the array still contained only the proxies
which defeated the indexOf to determine the next proposal/candidate.

Now the actual raw objects are added to the arrays.